Consolidate the per-head Muon tests into two files
@pengdurice asked for fewer files. Five, one per concern, become two split by
what they need to run:

  tests/unit/runtime/zero/test_per_head_muon.py            CPU
    the arithmetic          (was test_per_head_muon.py)
    which parameters are tagged, and with how many heads
                            (was test_per_head_muon_tagging.py)
    re-resolving the count against a tensor-parallel shard
                            (was test_per_head_muon_tensor_parallel.py)

  tests/unit/ops/muon/test_per_head_muon_accelerator.py    GPU
    a shard of the per-head result is the per-head result of the shard
                            (was test_per_head_muon_under_sharding.py)
    end-to-end training     (was test_per_head_muon_e2e.py)

Nothing is dropped or rewritten: 77 tests before, 77 after, same names. The
tagging module's _Attn and the tensor-parallel module's _Attn were different
classes with the same name, so the latter is now _ShardedAttn.

The accelerator file is not called test_per_head_muon.py because these
directories have no __init__.py, and two same-named modules break collection
when both are selected in one run.
